moved download modal integration tests to main.test because they depend on the BrowserWindow instance in main's scope. #136.
