+++ src/app.jsx
@@ -67,10 +67,15 @@ export default class App extends React.Component {
     });
 
     ipcRenderer.on('download-status', (event, downloadedNofN) => {
+      console.log('from renderer download-status')
       this.setState({
         downloadedNofN: downloadedNofN
       });
     });
+  }
+
+  componentWillUnmount() {
+    ipcRenderer.removeAllListeners('download-status');
   }
 
   /** Change the tab that is currently visible.
